refactor(middleware): move sandbox config under agent/middleware/, drop sandbox/ dir
- sandbox/config.{ts,test.ts} → agent/middleware/sandbox-config.{ts,test.ts}
  (renamed to avoid collision with the top-level src/config.ts)
- sandbox/ folder removed entirely
- All consumers (config.ts yaml loader, agent/runtime, agent/types, middleware)
  updated to the new path
