+++ src/agent/types.ts
@@ -1,4 +1,4 @@
-import type { SandboxConfig } from "../sandbox/config.js";
+import type { SandboxConfig } from "./middleware/sandbox-config.js";
 import type { AgentToolSettings } from "./tools/types.js";
 import type { DefaultSessionStore } from "./runners/pi/session-store.js";
